@@ -1,8 +1,7 @@
 module github.com/Jagiello95/toolbox
 
-go 1.17
+go 1.18
 
-require (
-	github.com/gabriel-vasile/mimetype v1.4.0 // indirect
-	golang.org/x/net v0.0.0-20210505024714-0287a6fb4125 // indirect
-)
+require github.com/gabriel-vasile/mimetype v1.4.0
+
+require golang.org/x/net v0.0.0-20210505024714-0287a6fb4125 // indirect
